Update JSArray extension to use `int` types. (#64)
This adapts the generator to work with latest Dart SDK. Recently the SDK
removed `JSNumber.toDart`. I could have replaced the calls with
`.toDartInt`, but instead, I switched the `JSArray.length` to specify an
`int` return value. In the process I also updated the indexing operator
to accept `int` values, and remove the need to explicitly convert
indices to a JSNumber as well.