Add generation code for Element constructors
For every tag that the interface supports, emits a Dart
constructor that can be used to create an element of that tag
using either createElement or createElementNS. Adds a dependency
to @webref/elements to fetch the element data.

Note that not every element interface has a tag associated with it.