Add a generate-all flag to emit all bindings
In the case where a user wants to use an experimental API
that we don't include in package:web, it is useful to have the
generator emit bindings still so they can copy it over instead
of having to write all of it themselves. This also allows us to
find problems in our tools with newer bindings before they
become standard.

- Adds a generate-all flag to the tools and refactors code to
incorporate it when determining whether to generate declarations.
- Cleans up argument passing so we're always using package:args.
- Adds a generate_all_and_analyze to the Dart CI.