Refactor generation logic so partials come after
Currently, the generator adds members to interfacelikes as we
process various IDL files. This may lead to partial interfaces
being added before the interface definition. Instead, this CL
adds partials after the interface definition, and then applies
mixins. It also dedups some logic around browser compatibility
so that we only check compatibility and process in one location.

While there is a diff in the generated files, it is functionally
the same because all we do is move members around in the same
extension type.