[interop] Add Support for Anonymous Declarations (#434)
This adds support for generating declarations for anonymous declarations, such as anonymous objects, unions, closures and constructor types.

This also adds support for nullable types (types unioned with undefined and/or null).

A hashing function is used for consistently hashing string objects, which is used for hashing identifiers for anonymous unions, objects and more, as well as comparing such names to allow reusing of such types.

* wip: union types

* wip: completed anonymous unions

* implemented JS Tuple

* completed nullability support for `undefined` and `null`

* wip: object decl

* implemented anonymous objects

* added support for closures and constructors

* wip: type hierarchy

* implemented sub type deduction

* added type generics to union types

* changed `.reduce` to `.fold`

* resolved newline and license headers

* isNullable updates and renamed `DeclarationAssociatedType`

* updated algorithm to use LCA via Topological Ordering

* resolved some more comments

* rm hasher_test

* resolved some more comments

* refactored tuple generation (common types) and more

* added LCA test

* removed stray prints

* added doc support and resolved merge

* updated documentation formatting using formatting.dart