[interop] Add support for Conditional and Predicate Types (#468)
Fixes #464 Fixes #463

This PR adds support for transforming conditional types (`t extends a ?
b : c`) and predicate types (`t is a`), and generates Dart alternatives
to these types.

Predicate types are converted to booleans, while conditional types are
converted to unions, since the result can either be one or the other.
There currently isn't a way (afaik) to introspect into parameter values
to give a specific result depending on the param result.

We could be smart, however, and wrap such methods with other methods to
do the casting for us in the case of conditional types, and could _try_
to do similar in predicate types (via `assert` checks and more).

Signed-off-by: Nike Okoronkwo <nikechukwu@gmail.com>