[interop] Add support for destructured parameters (#469)
* [interop] Added support for destructured parameters

Added support for working with destructured parameters, as well as other
kind of parameters. These take the place of the param type, since
destructuring in parameters is not allowed in Dart.

However, this goes a step further and provides documentation on the
parameter itself, providing the form of the param as documentation.

Signed-off-by: Nike Okoronkwo <nikechukwu@gmail.com>

* removed comma in comment

* added unknown case

---------

Signed-off-by: Nike Okoronkwo <nikechukwu@gmail.com>